STYLE: Prefer specifying b-values as `float` types
Prefer specifying b-values as `float` types consistently and avoid using
`numbers.Number` which seems to be problematic for type checking with
`mypy`:
https://github.com/python/mypy/issues/3186

+++ src/nifreeze/data/dmri.py
@@ -373,7 +373,7 @@ def find_shelling_scheme(
     bvals: np.ndarray,
     num_bins: int = DEFAULT_NUM_BINS,
     multishell_nonempty_bin_count_thr: int = DEFAULT_MULTISHELL_BIN_COUNT_THR,
-    bval_cap: int = DEFAULT_HIGHB_THRESHOLD,
+    bval_cap: float = DEFAULT_HIGHB_THRESHOLD,
 ) -> tuple[str, list[npt.NDArray[np.floating]], list[np.floating]]:
     """
     Find the shelling scheme on the given b-values.
@@ -391,7 +391,7 @@ def find_shelling_scheme(
         Number of bins.
     multishell_nonempty_bin_count_thr : :obj:`int`, optional
         Bin count to consider a multi-shell scheme.
-    bval_cap : :obj:`int`, optional
+    bval_cap : :obj:`float`, optional
         Maximum b-value to be considered in a multi-shell scheme.
 
     Returns
